File organization
## add:
- graphql folder under src
- schema.graphql under graphql
- resolvers.ts under graphql

## update:
- small change on index.ts
