+++ package.json
@@ -7,8 +7,8 @@
     "compile": "tsc",
     "start": "npm run compile && node ./dist/index.js",
     "dev": "tsx watch src/index.ts",
-    "check": "npx eslint-config-prettier src/index.ts",
-    "eslint": "npx eslint ."
+    "check": "npx eslint-config-prettier src/**/*.ts",
+    "eslint": "npx eslint src/**/*.ts"
   },
   "keywords": [
     "apollo",
